feat: update folders for frontend and put hash on backend
